# Index entries refactor and optimization