--- a/core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/OAbstractPaginatedStorage.java
+++ b/core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/OAbstractPaginatedStorage.java
@@ -2542,12 +2542,12 @@ public abstract class OAbstractPaginatedStorage extends OStorageAbstract
 
     for (OTransactionIndexChangesPerKey.OTransactionIndexEntry op :
         index.interpretTxKeyChanges(changes)) {
-      switch (op.operation) {
+      switch (op.getOperation()) {
         case PUT:
-          index.doPut(this, changes.key, op.value.getIdentity());
+          index.doPut(this, changes.key, op.getValue().getIdentity());
           break;
         case REMOVE:
-          if (op.value != null) index.doRemove(this, changes.key, op.value.getIdentity());
+          if (op.getValue() != null) index.doRemove(this, changes.key, op.getValue().getIdentity());
           else index.doRemove(this, changes.key);
           break;
         case CLEAR:
@@ -6604,7 +6604,7 @@ public abstract class OAbstractPaginatedStorage extends OStorageAbstract
           break;
         }
       }
-      if (!fullyLocked && !changes.nullKeyChanges.entries.isEmpty()) {
+      if (!fullyLocked && !changes.nullKeyChanges.isEmpty()) {
         index.acquireAtomicExclusiveLock(null);
       }
     }