--- a/core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/OMicroTransaction.java
+++ b/core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/OMicroTransaction.java
@@ -617,8 +617,8 @@ public final class OMicroTransaction implements OBasicTransaction, OTransactionI
     if (changesPerKey == null) return;
 
     for (final OTransactionIndexChangesPerKey.OTransactionIndexEntry indexEntry :
-        changesPerKey.entries)
-      if (indexEntry.value.getIdentity().equals(oldRid)) indexEntry.value = newRid;
+        changesPerKey.getEntriesAsList())
+      if (indexEntry.getValue().getIdentity().equals(oldRid)) indexEntry.setValue(newRid);
   }
 
   private static Dependency[] getIndexFieldRidDependencies(OIndex index) {