--- a/distributed/src/main/java/com/orientechnologies/orient/server/distributed/impl/task/OTransactionPhase1Task.java
+++ b/distributed/src/main/java/com/orientechnologies/orient/server/distributed/impl/task/OTransactionPhase1Task.java
@@ -393,7 +393,7 @@ public class OTransactionPhase1Task extends OAbstractReplicatedTask implements O
                 for (Object keyWithChange : changes.changesPerKey.keySet()) {
                   uniqueIndexKeys.add(new OTransactionUniqueKey(index, keyWithChange, 0));
                 }
-                if (!changes.nullKeyChanges.entries.isEmpty()) {
+                if (!changes.nullKeyChanges.isEmpty()) {
                   uniqueIndexKeys.add(new OTransactionUniqueKey(index, null, 0));
                 }
               }