--- a/lucene/src/main/java/com/orientechnologies/spatial/index/OLuceneSpatialIndex.java
+++ b/lucene/src/main/java/com/orientechnologies/spatial/index/OLuceneSpatialIndex.java
@@ -23,7 +23,6 @@ import com.orientechnologies.orient.core.tx.OTransactionIndexChanges;
 import com.orientechnologies.orient.core.tx.OTransactionIndexChangesPerKey;
 import com.orientechnologies.spatial.engine.OLuceneSpatialIndexContainer;
 import com.orientechnologies.spatial.shape.OShapeFactory;
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -74,7 +73,7 @@ public class OLuceneSpatialIndex extends OLuceneIndexNotUnique {
             if (((OLuceneSpatialIndexContainer) engine).isLegacy()) {
               return OLuceneSpatialIndex.super.interpretTxKeyChanges(changes);
             } else {
-              return interpretAsSpatial(changes.entries);
+              return interpretAsSpatial(changes);
             }
           });
     } catch (OInvalidIndexEngineIdException e) {
@@ -105,20 +104,19 @@ public class OLuceneSpatialIndex extends OLuceneIndexNotUnique {
   }
 
   private static Iterable<OTransactionIndexChangesPerKey.OTransactionIndexEntry> interpretAsSpatial(
-      List<OTransactionIndexChangesPerKey.OTransactionIndexEntry> entries) {
+      OTransactionIndexChangesPerKey item) {
     // 1. Handle common fast paths.
 
-    List<OTransactionIndexChangesPerKey.OTransactionIndexEntry> newChanges = new ArrayList<>();
-
+    List<OTransactionIndexChangesPerKey.OTransactionIndexEntry> entries = item.getEntriesAsList();
     Map<OIdentifiable, Integer> counters = new LinkedHashMap<>();
 
     for (OTransactionIndexChangesPerKey.OTransactionIndexEntry entry : entries) {
 
-      Integer counter = counters.get(entry.value);
+      Integer counter = counters.get(entry.getValue());
       if (counter == null) {
         counter = 0;
       }
-      switch (entry.operation) {
+      switch (entry.getOperation()) {
         case PUT:
           counter++;
           break;
@@ -128,24 +126,22 @@ public class OLuceneSpatialIndex extends OLuceneIndexNotUnique {
         case CLEAR:
           break;
       }
-      counters.put(entry.value, counter);
+      counters.put(entry.getValue(), counter);
     }
+
+    OTransactionIndexChangesPerKey changes = new OTransactionIndexChangesPerKey(item.key);
 
     for (Map.Entry<OIdentifiable, Integer> entry : counters.entrySet()) {
       OIdentifiable oIdentifiable = entry.getKey();
       switch (entry.getValue()) {
         case 1:
-          newChanges.add(
-              new OTransactionIndexChangesPerKey.OTransactionIndexEntry(
-                  oIdentifiable, OTransactionIndexChanges.OPERATION.PUT));
+          changes.add(oIdentifiable, OTransactionIndexChanges.OPERATION.PUT);
           break;
         case -1:
-          newChanges.add(
-              new OTransactionIndexChangesPerKey.OTransactionIndexEntry(
-                  oIdentifiable, OTransactionIndexChanges.OPERATION.REMOVE));
+          changes.add(oIdentifiable, OTransactionIndexChanges.OPERATION.REMOVE);
           break;
       }
     }
-    return newChanges;
+    return changes.getEntriesAsList();
   }
 }