Refactor of OTransactionIndexChangesPerKey: all the fields were public, accessed in read/write mode everywhere, impossible to optimise without breaking things. 
This refactor reduces visibility of all the fields and encapsulates the get/set behaviour to allow further optimizations.

Second step is the optimization of OTransactionIndexChangesPerKey.add(), to avoid iteration of all the entries at each execution (O(N^2) on the tx size)