@luigidellaquila 2021-05-06
Some microbenchmarking:

This fix adds a 10% overhead to small transactions and in cases where there are no collisions on the keys.

On a case with many collisions, this implementation is constant time (ie. same performance as no collisions), while the previous one tends to increase the execution time.

eg.

 100 tx, 20k records per tx, 100% key collisions (ie. all the records have the same index key)
- old implementation takes ~200 seconds
- new implementation takes ~50 seconds

 100 tx, 20k records per tx, 10 different keys equally distributed in the record set 
- old implementation takes ~75 seconds
- new implementation takes ~50 seconds

 100 tx, 20k records per tx, 100 different keys equally distributed in the record set 
- old implementation takes ~50 seconds
- new implementation takes ~50 seconds